Proyecto final IaStock/requirements.txt
django>=4.2
djangorestframework
celery
redis
pandas
scikit-learn
requests
